Move the version status column from default set to available column
