+++ concrete/src/Page/Search/ColumnSet/DefaultSet.php
@@ -5,7 +5,6 @@ use Concrete\Core\Page\Search\ColumnSet\Column\CollectionVersionColumn;
 use Concrete\Core\Page\Search\ColumnSet\Column\DateLastModifiedColumn;
 use Concrete\Core\Page\Search\ColumnSet\Column\DatePublicColumn;
 use Concrete\Core\Search\Column\Column;
-use Concrete\Core\Support\Facade\Application;
 use UserInfo;
 use Core;
 
@@ -32,33 +31,6 @@ class DefaultSet extends ColumnSet
         }
     }
 
-    /**
-     * @see \Concrete\Core\Page\Collection\Version\Version::get()
-     *
-     * @param \Concrete\Core\Page\Page $c
-     *
-     * @return string
-     */
-    public static function getCollectionVersionStatus($c)
-    {
-        $cvStatus = '';
-        $app = Application::getFacadeApplication();
-        $now = $app->make('date')->getOverridableNow();
-
-        $vObj = $c->getVersionObject();
-        if ($vObj) {
-            if ($vObj->isApproved() && (!$vObj->getPublishDate() || $vObj->getPublishDate() <= $now) && (!$vObj->getPublishEndDate() || $vObj->getPublishEndDate() >= $now)) {
-                $cvStatus = t('Approved');
-            } elseif ($vObj->isApproved() && ($vObj->getPublishDate() && $vObj->getPublishDate() > $now)) {
-                $cvStatus = t('Scheduled');
-            } elseif (!$vObj->isApproved()) {
-                $cvStatus = t('Unapproved');
-            }
-        }
-
-        return $cvStatus;
-    }
-
     public function __construct()
     {
         $this->addColumn(new Column('pt.ptHandle', t('Type'), 'getPageTypeName', false));
@@ -66,7 +38,6 @@ class DefaultSet extends ColumnSet
         $this->addColumn(new DatePublicColumn());
         $this->addColumn(new DateLastModifiedColumn());
         $this->addColumn(new Column('author', t('Author'), ['\Concrete\Core\Page\Search\ColumnSet\DefaultSet', 'getCollectionAuthor'], false));
-        $this->addColumn(new Column('cvStatus', t('Version Status'), ['\Concrete\Core\Page\Search\ColumnSet\DefaultSet', 'getCollectionVersionStatus'], false));
         $date = $this->getColumnByKey('c.cDateModified');
         $this->setDefaultSortColumn($date, 'desc');
     }
